Move object_store and volume_backup manual tests from test_manual.py
